requirements.txt
flask>=3.0.0
python-whois>=0.9.4
gunicorn>=21.0.0
